+++ src/CacheControl/CacheControlServiceProvider.php
@@ -2,6 +2,7 @@
 
 namespace Nuwave\Lighthouse\CacheControl;
 
+use GraphQL\Language\AST\DirectiveNode;
 use GraphQL\Language\AST\FieldNode;
 use GraphQL\Language\AST\NodeKind;
 use GraphQL\Language\Visitor;
@@ -43,15 +44,7 @@ class CacheControlServiceProvider extends ServiceProvider
                         ? ASTHelper::directiveDefinition($field->astNode, 'cacheControl')
                         : null;
                     if ($cacheControlDirective !== null) {
-                        if (! ASTHelper::directiveArgValue($cacheControlDirective, 'inheritMaxAge')) {
-                            $cacheControl->addMaxAge(
-                                ASTHelper::directiveArgValue($cacheControlDirective, 'maxAge') ?? 0,
-                            );
-                        }
-
-                        if (ASTHelper::directiveArgValue($cacheControlDirective, 'scope') === 'PRIVATE') {
-                            $cacheControl->setPrivate();
-                        }
+                        self::setCacheValues($cacheControlDirective, $cacheControl);
                     } else {
                         $parent = $typeInfo->getParentType();
                         assert($parent instanceof NamedType);
@@ -65,7 +58,12 @@ class CacheControlServiceProvider extends ServiceProvider
                                 $nodeType = $nodeType->getInnermostType();
                             }
 
-                            if (! $nodeType instanceof ScalarType) {
+                            $cacheControlDirective = isset($nodeType->astNode)
+                                ? ASTHelper::directiveDefinition($nodeType->astNode, 'cacheControl')
+                                : null;
+                            if ($cacheControlDirective !== null) {
+                                self::setCacheValues($cacheControlDirective, $cacheControl);
+                            } elseif (! $nodeType instanceof ScalarType) {
                                 $cacheControl->addMaxAge(0);
                             }
                         }
@@ -91,5 +89,19 @@ class CacheControlServiceProvider extends ServiceProvider
 
             $this->app->forgetInstance(CacheControl::class);
         });
+    }
+
+    /** Set HTTP cache header values based on the @cacheControl directive. */
+    private static function setCacheValues(DirectiveNode $cacheControlDirective, CacheControl $cacheControl): void
+    {
+        if (! ASTHelper::directiveArgValue($cacheControlDirective, 'inheritMaxAge')) {
+            $cacheControl->addMaxAge(
+                ASTHelper::directiveArgValue($cacheControlDirective, 'maxAge') ?? 0,
+            );
+        }
+
+        if (ASTHelper::directiveArgValue($cacheControlDirective, 'scope') === 'PRIVATE') {
+            $cacheControl->setPrivate();
+        }
     }
 }