+++ tests/Integration/CacheControl/CacheControlDirectiveTest.php
@@ -312,4 +312,106 @@ final class CacheControlDirectiveTest extends DBTestCase
             ],
         ];
     }
+
+    public function testUsePaginate(): void
+    {
+        $this->schema = /** @lang GraphQL */ '
+        type User {
+            tasks: [Task!]! @hasMany(type: PAGINATOR) @cacheControl(maxAge: 50)
+        }
+        
+        type Task {
+            id: Int @cacheControl(maxAge: 10)
+            foo: String @cacheControl(inheritMaxAge: true)
+            bar: String
+        }
+
+        type Query {
+            users: [User] @paginate @cacheControl(maxAge: 5, scope: PRIVATE)
+        }
+        ';
+
+        $user = factory(User::class)->create();
+        assert($user instanceof User);
+
+        $tasks = factory(Task::class, 3)->make();
+        $user->tasks()->saveMany($tasks);
+
+        $this->graphQL(/** @lang GraphQL */ '
+                {
+                    users(first:10) {
+                        paginatorInfo {
+                            count
+                        }
+                       data {
+                            tasks(first:10) {
+                                data {
+                                    id
+                                    foo
+                                }
+                            }
+                       }    
+                    }
+                }
+            ')
+            ->assertHeader('Cache-Control', 'max-age=5, private');
+    }
+
+    /** @dataProvider typeLevelCacheDataProvider */
+    public function testTypeLevelCache(string $query, string $expectedHeaderString): void
+    {
+        $this->schema = /** @lang GraphQL */ '
+        type User {
+            tasks: [Task!]! @hasMany
+            tasksWithCache: [Task!]! @hasMany(relation: "tasks") @cacheControl(maxAge: 20)
+        }
+        
+        type Task @cacheControl(maxAge: 10){
+            id: Int 
+            foo: String
+            bar: String
+        }
+
+        type Query {
+            user: User @first @cacheControl(maxAge: 50, scope: PRIVATE)
+        }
+        ';
+
+        $user = factory(User::class)->create();
+        assert($user instanceof User);
+
+        $tasks = factory(Task::class, 3)->make();
+        $user->tasks()->saveMany($tasks);
+
+        $this->graphQL($query)
+            ->assertHeader('Cache-Control', $expectedHeaderString);
+    }
+
+    /** @return iterable<array{string, string}> */
+    public static function typeLevelCacheDataProvider(): iterable
+    {
+        yield [/** @lang GraphQL */ '
+                {
+                    user {
+                        tasks {
+                            id
+                            foo
+                        }
+                    }
+                }
+            ', 'max-age=10, private',
+        ];
+
+        yield [/** @lang GraphQL */ '
+                {
+                    user {
+                        tasksWithCache {
+                            id
+                            foo
+                        }
+                    }
+                }
+            ', 'max-age=20, private',
+        ];
+    }
 }