+++ CHANGELOG.md
@@ -9,6 +9,14 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 ## Unreleased
 
+### Added
+
+- Allow `@cacheControl` directive on type level https://github.com/nuwave/lighthouse/pull/2451
+
+### Fixed
+
+- Inherit `@cacheControl` in pagination on fields https://github.com/nuwave/lighthouse/pull/2451
+
 ## v6.19.0
 
 ### Added